requirements.txt
airtest==1.3.1
Appium_Python_Client==3.1.0
poco==0.99.1
pocoui==1.0.92
retrying
selenium
unittestreport
MarkupSafe==2.0.1
